Squashed commit of the following:
commit fca901e153acc60155951d02a7d16e8aebfb513b
Author: devTap <rlaxogus0517@gmail.com>
Date:   Wed Apr 12 22:15:55 2023 +0900

    helper 추가 + onCreate 수정
